// src/app/core/model/special_import_license_info.ts
export interface SpecialImportLicenseInfo {
  licenseNo: string;
  formNo: string;
  scFlagType: string;
  ouCode: string;
  ouName: string;
  licenseType: string;
  descript: string;
  productCode: string;
  usedQuantity: number;
  sc047UsedQuantity: number;
  controlQty: number;
  startDate: number;
  endDate: number;
}